fix: correct SW registration root-cause docs, split unavailable-workbox guard, add idempotency (WA-2949)
Review round 1 follow-up:
- Correct the root-cause description: the production updatefound TypeError
  is caused by register() REJECTING because workbox-window's internal
  registration value is truthy but not a real ServiceWorkerRegistration
  (a WebView stub lacking addEventListener), not by register() resolving
  undefined. Fixed the module doc and the misleadingly-named test; the
  now-inert 'resolves without a registration object' test (couldn't fail
  either way — an async void function resolves undefined whether the
  guard exists or not) is replaced by a warn-path test and an idempotency
  test that are actually revert-sensitive.
- Split the single 'window unavailable' guard: serviceWorker unsupported
  stays a silent no-op, but serviceWorker supported + window.workbox
  missing (should be impossible) now logs a warning instead of failing
  silently — that combination would otherwise mean zero telemetry for a
  total SW registration outage.
- Added a module-level idempotency guard so calling registerServiceWorker()
  more than once is a safe no-op after the first attempt.
- Moved test files to match the Test Decision Matrix in apps/web/AGENTS.md:
  hook tests in hooks/__tests__/, service tests colocated.
