+++ apps/web/src/services/pwa/registerServiceWorker.test.ts
@@ -1,4 +1,4 @@
-import { registerServiceWorker } from '../registerServiceWorker'
+import { registerServiceWorker, __resetRegisterServiceWorkerForTests } from './registerServiceWorker'
 
 jest.mock('@/services/observability', () => ({
   logger: {
@@ -9,13 +9,14 @@ jest.mock('@/services/observability', () => ({
   },
 }))
 
-const { logger } = jest.requireMock('@/services/observability') as { logger: { info: jest.Mock } }
+const { logger } = jest.requireMock('@/services/observability') as { logger: { info: jest.Mock; warn: jest.Mock } }
 
 describe('registerServiceWorker', () => {
   const originalServiceWorker = (navigator as unknown as { serviceWorker?: unknown }).serviceWorker
 
   afterEach(() => {
     jest.clearAllMocks()
+    __resetRegisterServiceWorkerForTests()
     delete (window as unknown as { workbox?: unknown }).workbox
     Object.defineProperty(navigator, 'serviceWorker', {
       value: originalServiceWorker,
@@ -45,6 +46,7 @@ describe('registerServiceWorker', () => {
 
     expect(register).toHaveBeenCalledTimes(1)
     expect(logger.info).not.toHaveBeenCalled()
+    expect(logger.warn).not.toHaveBeenCalled()
   })
 
   it('does nothing when serviceWorker is not supported (e.g. some private-browsing modes)', async () => {
@@ -57,19 +59,17 @@ describe('registerServiceWorker', () => {
     expect(register).not.toHaveBeenCalled()
   })
 
-  it('does nothing when window.workbox was never created', async () => {
+  it('logs a warning and does nothing when serviceWorker is supported but window.workbox is unavailable', async () => {
+    // Should be impossible in practice — next-pwa's injected script always
+    // assigns `window.workbox` whenever `serviceWorker` is supported. This
+    // guards against that invariant silently breaking (e.g. a next-pwa
+    // version bump), which would otherwise stop SW registration for every
+    // user with zero telemetry.
     setServiceWorkerSupport(true)
 
     await expect(registerServiceWorker()).resolves.toBeUndefined()
-  })
 
-  it('does not throw and does not log when registration resolves without a registration object', async () => {
-    setServiceWorkerSupport(true)
-    const register = jest.fn().mockResolvedValue(undefined)
-    window.workbox = { register } as unknown as Window['workbox']
-
-    await expect(registerServiceWorker()).resolves.toBeUndefined()
-    expect(logger.info).not.toHaveBeenCalled()
+    expect(logger.warn).toHaveBeenCalledWith('Service worker registration skipped: window.workbox is unavailable')
   })
 
   it('catches a rejected registration (e.g. AbortError on unreachable sw.js) and logs it at info level', async () => {
@@ -84,7 +84,13 @@ describe('registerServiceWorker', () => {
     expect(logger.info).toHaveBeenCalledWith('Service worker registration failed', { error: abortError.message })
   })
 
-  it('catches the TypeError thrown when workbox-window attaches "updatefound" to an undefined registration', async () => {
+  it('catches the TypeError workbox-window throws when the resolved value is not a real ServiceWorkerRegistration (WebView stub) and logs it at info level', async () => {
+    // This is the actual shape of the production `updatefound` cluster:
+    // `navigator.serviceWorker.register()` resolves with something truthy
+    // (some in-app WebViews return a stub), so workbox-window's own
+    // `.waiting` read doesn't throw — but the stub has no `addEventListener`,
+    // so workbox-window's later `registration.addEventListener('updatefound', ...)`
+    // throws, rejecting the promise `window.workbox.register()` returns.
     setServiceWorkerSupport(true)
     const typeError = new TypeError('l.fn.addEventListener is not a function')
     const register = jest.fn().mockRejectedValue(typeError)
@@ -93,5 +99,17 @@ describe('registerServiceWorker', () => {
     await expect(registerServiceWorker()).resolves.toBeUndefined()
 
     expect(logger.info).toHaveBeenCalledWith('Service worker registration failed', { error: typeError.message })
+  })
+
+  it('only attempts registration once even if called multiple times', async () => {
+    setServiceWorkerSupport(true)
+    const register = jest.fn().mockResolvedValue({ scope: '/' })
+    window.workbox = { register } as unknown as Window['workbox']
+
+    await registerServiceWorker()
+    await registerServiceWorker()
+    await registerServiceWorker()
+
+    expect(register).toHaveBeenCalledTimes(1)
   })
 })
